Skapat och flyttat modeller till separat projekt
