+++ ResuMate.Api/Controllers/CvController.cs
@@ -29,6 +29,8 @@ public class CvController : ControllerBase
         return Ok(new { message = $"Hello, {name}! Welcome to our API." });
     }
 
+
+
 }
 
 public class TestRequest
