+++ ResuMate.Api/ResuMate.Api.csproj
@@ -13,4 +13,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="7.3.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\ResuMate.Shared\ResuMate.Shared.csproj" />
+  </ItemGroup>
+
 </Project>
