+++ ResuMate.Shared/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace ResuMate.Shared
+{
+    public class Class1
+    {
+
+    }
+}
