+++ ResuMate.Shared/Models/CvTemplate.cs
@@ -1,0 +1,9 @@
+﻿namespace ResuMate.Shared.Models
+{
+    public enum CvTemplate
+    {
+        Classic,
+        Modern,
+        Professional,       
+    }
+}
