+++ ResuMate.Shared/Models/EducationModel.cs
@@ -1,0 +1,12 @@
+namespace ResuMate.Shared.Models
+{
+    public class EducationModel
+    {
+        public string? School { get; set; }
+        public string? Degree { get; set; }
+        public string? Specialization { get; set; }
+        public string? StartYear { get; set; }
+        public string? EndYear { get; set; }
+
+    }
+}
