+++ ResuMate.Shared/Models/ExperienceModel.cs
@@ -1,0 +1,11 @@
+namespace ResuMate.Shared.Models
+{
+    public class ExperienceModel
+    {
+        public string? JobTitle { get; set; }
+        public string? Company { get; set; }
+        public string? JobDescription { get; set; }
+        public string? StartYear { get; set; }
+        public string? EndYear { get; set; }
+    }
+}
