+++ ResuMate.Shared/Models/ReferenceModel.cs
@@ -1,0 +1,9 @@
+﻿namespace ResuMate.Shared.Models
+{
+    public class ReferenceModel
+    {
+        public string? Name { get; set; } = string.Empty;
+        public string? Relation { get; set; } = string.Empty;
+        public string? PhoneNumber { get; set; } = string.Empty;
+    }
+}
