+++ ResuMate.sln
@@ -4,11 +4,16 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 17.13.35825.156
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ResuMate", "ResuMate\ResuMate.csproj", "{EBD6EE73-2F15-4179-A794-5812155EB674}"
+	ProjectSection(ProjectDependencies) = postProject
+		{BF69BDA5-347F-49DF-94D6-04714F3F9827} = {BF69BDA5-347F-49DF-94D6-04714F3F9827}
+	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ResuMate.Api", "ResuMate.Api\ResuMate.Api.csproj", "{8F29D7AD-0300-46D0-BBE1-18236C00FAE4}"
 	ProjectSection(ProjectDependencies) = postProject
-		{EBD6EE73-2F15-4179-A794-5812155EB674} = {EBD6EE73-2F15-4179-A794-5812155EB674}
+		{BF69BDA5-347F-49DF-94D6-04714F3F9827} = {BF69BDA5-347F-49DF-94D6-04714F3F9827}
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ResuMate.Shared", "ResuMate.Shared\ResuMate.Shared.csproj", "{BF69BDA5-347F-49DF-94D6-04714F3F9827}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -24,6 +29,10 @@ Global
 		{8F29D7AD-0300-46D0-BBE1-18236C00FAE4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8F29D7AD-0300-46D0-BBE1-18236C00FAE4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8F29D7AD-0300-46D0-BBE1-18236C00FAE4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BF69BDA5-347F-49DF-94D6-04714F3F9827}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BF69BDA5-347F-49DF-94D6-04714F3F9827}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BF69BDA5-347F-49DF-94D6-04714F3F9827}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BF69BDA5-347F-49DF-94D6-04714F3F9827}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
