+++ ResuMate/Components/Models/EducationModel.cs
@@ -1,12 +1,12 @@
-namespace ResuMate.Components.Models
-{
-    public class EducationModel
-    {
-        public string? School { get; set; }
-        public string? Degree { get; set; }
-        public string? Specialization { get; set; }
-        public string? StartYear { get; set; }
-        public string? EndYear { get; set; }
+//namespace ResuMate.Components.Models
+//{
+//    public class EducationModel
+//    {
+//        public string? School { get; set; }
+//        public string? Degree { get; set; }
+//        public string? Specialization { get; set; }
+//        public string? StartYear { get; set; }
+//        public string? EndYear { get; set; }
 
-    }
-}
+//    }
+//}
