+++ ResuMate/Components/Models/ExperienceModel.cs
@@ -1,11 +1,11 @@
-namespace ResuMate.Components.Models
-{
-    public class ExperienceModel
-    {
-        public string? JobTitle { get; set; }
-        public string? Company { get; set; }
-        public string? JobDescription { get; set; }
-        public string? StartYear { get; set; }
-        public string? EndYear { get; set; }
-    }
-}
+//namespace ResuMate.Components.Models
+//{
+//    public class ExperienceModel
+//    {
+//        public string? JobTitle { get; set; }
+//        public string? Company { get; set; }
+//        public string? JobDescription { get; set; }
+//        public string? StartYear { get; set; }
+//        public string? EndYear { get; set; }
+//    }
+//}
