+++ ResuMate/Components/Models/ReferenceModel.cs
@@ -1,9 +1,9 @@
-﻿namespace ResuMate.Components.Models
-{
-    public class ReferenceModel
-    {
-        public string? Name { get; set; } = string.Empty;
-        public string? Relation { get; set; } = string.Empty;
-        public string? PhoneNumber { get; set; } = string.Empty;
-    }
-}
+﻿//namespace ResuMate.Components.Models
+//{
+//    public class ReferenceModel
+//    {
+//        public string? Name { get; set; } = string.Empty;
+//        public string? Relation { get; set; } = string.Empty;
+//        public string? PhoneNumber { get; set; } = string.Empty;
+//    }
+//}
