+++ ResuMate/Components/Pages/CvForm.razor
@@ -2,6 +2,7 @@
 @using QuestPDF.Fluent
 @using ResuMate.Services
 @using ResuMate.Components.Models
+@using ResuMate.Shared.Models
 @rendermode InteractiveServer
 @inject CreatePdfService CreatePdfService
