+++ ResuMate/Components/Pages/PersonalLetter.razor
@@ -1,8 +1,8 @@
 ﻿@page "/Personal-letter"
 @inject HttpClient Http
 @using QuestPDF.Fluent
-@using ResuMate.Services
 @using ResuMate.Components.Models
+@using ResuMate.Services
 @rendermode InteractiveServer
 
 <h1 class="section-title">Här kan du med hjälp av AI skriva ett personligt brev riktat till ditt arbete</h1>
