+++ ResuMate/ResuMate.csproj
@@ -17,6 +17,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ResuMate.Api\ResuMate.Api.csproj" />
+    <ProjectReference Include="..\ResuMate.Shared\ResuMate.Shared.csproj" />
   </ItemGroup>
 
 </Project>
